requirements.txt
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pymysql>=1.0.0
cryptography>=3.4.0
pydantic>=2.0.0

Flask>=2.3.3
flask-session2~=1.3.1
flask-socketio>=5.3.0
python-socketio>=5.10.0
openai>=1.50.0
plotly>=5.18.0